Filter menu name out of menu classes.
For reasons unknown, each word of the menu name appears as an
individual class in the list of classes associated with the
menu item. These were being concatenated together with hyphens,
but could end up with class names like 'course-content' which
has a special CSS meaning. For now we just filter out all
classes that don't contain a hyphen or underscore.

Workaround for #37